+++ packages/blockchain-utils-validation/src/index.ts
@@ -1,0 +1,35 @@
+import { LinkProof } from "@ceramicnetwork/blockchain-utils-linking";
+import ethereum from "./blockchains/ethereum";
+import filecoin from "./blockchains/filecoin";
+import polkadot from "./blockchains/polkadot";
+import eosio from "./blockchains/eosio";
+import { AccountID } from "caip";
+
+const handlers = {
+  [ethereum.namespace]: ethereum,
+  [filecoin.namespace]: filecoin,
+  [polkadot.namespace]: polkadot,
+  [eosio.namespace]: eosio,
+};
+
+const findDID = (did: string): string | undefined => did.match(/(did:[a-zA-Z0-9]+:[a-zA-Z0-9]+)/)?.[0]
+
+export async function validateLink(
+  proof: LinkProof
+): Promise<LinkProof | null> {
+  // version < 2 are always eip155 namespace
+  let namespace = ethereum.namespace;
+  if (proof.version >= 2) {
+    namespace = new AccountID(proof.account).chainId.namespace;
+  }
+  const handler = handlers[namespace];
+  if (!handler)
+    throw new Error(`proof with namespace '${namespace}' not supported`);
+  const validProof = await handler.validateLink(proof);
+  if (validProof) {
+    validProof.did = findDID(validProof.message);
+    return validProof;
+  } else {
+    return null;
+  }
+}