+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@0.19.2...@ceramicnetwork/cli@0.19.3) (2021-01-14)
+
+
+### Bug Fixes
+
+* send http error response as json ([#790](https://github.com/ceramicnetwork/js-ceramic/issues/790)) ([02e1dfc](https://github.com/ceramicnetwork/js-ceramic/commit/02e1dfcbe00a8508a6a2c5035b23156abbe723b8)), closes [#789](https://github.com/ceramicnetwork/js-ceramic/issues/789)
+
+
+
+
+
+## [0.19.2](/compare/@ceramicnetwork/cli@0.19.1...@ceramicnetwork/cli@0.19.2) (2021-01-13)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+## [0.19.1](/compare/@ceramicnetwork/cli@0.19.0...@ceramicnetwork/cli@0.19.1) (2021-01-13)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [0.19.0](/compare/@ceramicnetwork/cli@0.18.6...@ceramicnetwork/cli@0.19.0) (2021-01-13)
+
+
+### Bug Fixes
+
+* **cli,http-client:** Fix pin API in CLI and http client (#752) 20fcd75, closes #752
+
+
+### Features
+
+* **cli:** Allow specifying pub/sub topic for 'local' ceramic network (#781) f3650b4, closes #781
+
+
+
+
+
 ## [0.18.6](/compare/@ceramicnetwork/cli@0.18.5...@ceramicnetwork/cli@0.18.6) (2021-01-07)
 
 **Note:** Version bump only for package @ceramicnetwork/cli