+++ packages/pinning-powergate-backend/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.12](/compare/@ceramicnetwork/pinning-powergate-backend@0.1.11...@ceramicnetwork/pinning-powergate-backend@0.1.12) (2021-01-13)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-powergate-backend
+
+
+
+
+
+## [0.1.11](/compare/@ceramicnetwork/pinning-powergate-backend@0.1.10...@ceramicnetwork/pinning-powergate-backend@0.1.11) (2021-01-13)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-powergate-backend
+
+
+
+
+
+## [0.1.10](/compare/@ceramicnetwork/pinning-powergate-backend@0.1.9...@ceramicnetwork/pinning-powergate-backend@0.1.10) (2021-01-13)
+
+**Note:** Version bump only for package @ceramicnetwork/pinning-powergate-backend
+
+
+
+
+
 ## [0.1.9](/compare/@ceramicnetwork/pinning-powergate-backend@0.1.8...@ceramicnetwork/pinning-powergate-backend@0.1.9) (2021-01-07)
 
 **Note:** Version bump only for package @ceramicnetwork/pinning-powergate-backend